adding PS1 to bash_profile, adding iterm2 colors, fixing symlink issue

+++ link.py
@@ -20,7 +20,7 @@ if __name__ == '__main__':
         if not args.dry:
           symlink(dfpath, lnpath)
         print dfpath, ' >> ', lnpath
-      elif args.action == 'clean':
+      elif args.action == 'clean' and isfile(lnpath):
         if not args.dry:
           unlink(lnpath)
         print '!!', lnpath
